Refactored out additional Ionic references, combined SignUpStep and the SignUp page
