+++ src/components/InputWrapper/responsive.scss
@@ -2,7 +2,7 @@
   @media screen and (min-width: 600px) {
     margin: 48px 0 30px;
   }
-  ion-input {
+  input {
     @media screen and (min-width: 576px) {
       --padding-start: 18px;
     }
@@ -10,7 +10,7 @@
 }
 .plt-android {
   .input-wrapper {
-    ion-label {
+    label {
       @media screen and (min-width: 600px) {
         padding-bottom: 21px;
       }
