+++ src/pages/PrivacyPolicy/index.scss
@@ -1,6 +1,6 @@
 #welcome-section {
   .privacy-page {
-    ion-content {
+    .content {
       --background: white;
       h3 {
         font-size: 22px;
