+++ src/pages/TermsOfUse/index.scss
@@ -1,6 +1,6 @@
 #welcome-section {
   .terms-page {
-    ion-content {
+    .content {
       --background: white;
       h3 {
         font-size: 22px;
